pyproject.toml
[project]
name = "mitm-toolkit"
version = "0.1.0"
description = "A comprehensive toolkit for capturing and analyzing HTTP/HTTPS traffic to create local service replicas"
authors = [{name = "MITM Toolkit Contributors"}]
readme = "README.md"
license = "MIT"
repository = "https://github.com/binbandit/mitm-toolkit"
requires-python = ">=3.10"
dependencies = [
    "mitmproxy>=10.1.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "fastapi>=0.105.0",
    "uvicorn>=0.25.0",
    "httpx>=0.25.0",
    "jinja2>=3.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiofiles>=23.2.0",
    "websockets>=12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
mitm-toolkit = "mitm_toolkit.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]

[tool.mypy]
python_version = "3.10"
strict = true